+++ src/app/[variants]/(main)/discover/(list)/assistant/AssistantPage.tsx
@@ -13,7 +13,7 @@ import List from './features/List';
 import Loading from './loading';
 
 const AssistantPage = memo<{ mobile?: boolean }>(() => {
-  const { q, page, category, sort, order } = useQuery() as AssistantQueryParams;
+  const { q, page, category, sort, order, source } = useQuery() as AssistantQueryParams;
   const useAssistantList = useDiscoverStore((s) => s.useAssistantList);
   const { data, isLoading } = useAssistantList({
     category,
@@ -22,7 +22,10 @@ const AssistantPage = memo<{ mobile?: boolean }>(() => {
     pageSize: 21,
     q,
     sort,
+    source,
   });
+
+  console.log("data",data)
 
   if (isLoading || !data) return <Loading />;